+++ CHANGELOG.rst
@@ -4,6 +4,10 @@ Changelog
 1.2.5 (unreleased)
 ++++++++++++++++++
 
+Bug fixes:
+
+- Fix validation of invalid types passed to a ``Nested`` field when ``many=True`` (:issue:`188`).
+
 Support:
 
 - Fix pep8 dev dependency for flake8. Thanks :user:`taion`.